+++ src/Core/Core.Tests/CodeFirstTests.cs
@@ -1,6 +1,5 @@
 ﻿using System.Collections.Generic;
 using System.Threading.Tasks;
-using ChilliCream.Testing;
 using HotChocolate.Execution;
 using HotChocolate.Resolvers;
 using HotChocolate.Types;