+++ src/Core/Core.Tests/Execution/Instrumentation/ApolloTracingResultBuilderTests.cs
@@ -1,5 +1,4 @@
 ﻿using System;
-using ChilliCream.Testing;
 using Snapshooter.Xunit;
 using Xunit;