+++ src/Core/Types.Tests/Types/Descriptors/ObjectTypeDescriptorTests.cs
@@ -1,7 +1,6 @@
 ﻿using System;
 using System.Linq;
 using System.Threading.Tasks;
-using ChilliCream.Testing;
 using HotChocolate.Execution;
 using HotChocolate.Resolvers;
 using Snapshooter.Xunit;